Add domMocks test utilities and update tests
Extract repeated JSDOM/window mock setup into a new test utility (src/vertex/test/utils/domMocks.ts) exposing setupPointerEventMock and setupResizeObserverMock. Replace inline MockPointerEvent/ResizeObserver and HTMLElement overrides across multiple UI tests (chart, combobox, date-picker, dialog, dropdown-menu, multi-select, popover, select) with imports and calls to the new utilities. The helpers install mocks in beforeAll and restore originals in afterAll to keep tests isolated and reduce duplication.
