+++ src/vertex/components/ui/select.test.tsx
@@ -9,23 +9,11 @@ import {
   SelectValue,
 } from "./select";
 
-describe("Select", () => {
-  class MockPointerEvent extends Event {
-    button: number;
-    ctrlKey: boolean;
-    pointerType: string;
+import { setupPointerEventMock } from "@/test/utils/domMocks";
 
-    constructor(type: string, props: PointerEventInit) {
-      super(type, props);
-      this.button = props.button || 0;
-      this.ctrlKey = props.ctrlKey || false;
-      this.pointerType = props.pointerType || "mouse";
-    }
-  }
-  window.PointerEvent = MockPointerEvent as never;
-  window.HTMLElement.prototype.scrollIntoView = () => {};
-  window.HTMLElement.prototype.hasPointerCapture = () => false;
-  window.HTMLElement.prototype.releasePointerCapture = () => {};
+setupPointerEventMock();
+
+describe("Select", () => {
 
   it("renders and selects options", async () => {
     const handleValueChange = vi.fn();
